+++ src/components/EditLabelButton/EditLabelButton.tsx
@@ -1,13 +1,15 @@
 import type React from 'react';
 import { useCallback, useState } from 'react';
 import * as TEXTS from '../../constants/texts';
+import { getElementLabelTipProps } from '../../helpers/common';
 import * as gtag from '../../helpers/gtag';
 import type { ILabel } from '../../models/Label';
-import { actions as personalActions } from '../../store/slices/personal';
+import { actions as personalActions, IEditLabelPayload } from '../../store/slices/personal';
 import { useTypedDispatch } from '../../store/store';
 import { EventAction, EventCategory, EventLabel } from '../../types/ga';
 import { IconName } from '../Icon/types';
 import IconButton from '../IconButton/IconButton';
+import useLabelForm from '../LabelForm/useLabelForm';
 import LabelFormModal, { TLabelFormProps } from '../LabelFormModal/LabelFormModal';
 
 interface IProps {
@@ -21,10 +23,11 @@ type TComponentProps = React.ComponentPropsWithoutRef<'button'>;
 type TProps = IProps & TComponentProps;
 
 const EditLabelButton: React.FunctionComponent<TProps> = (props) => {
-  const { className, user, index, label } = props;
+  const { user, index, label, ...otherProps } = props;
 
   const dispatch = useTypedDispatch();
   const [open, setOpen] = useState(false);
+  const [loading, submit] = useLabelForm();
 
   const handleClick: React.MouseEventHandler<HTMLButtonElement> = useCallback((event) => {
     event.preventDefault();
@@ -41,21 +44,21 @@ const EditLabelButton: React.FunctionComponent<TProps> = (props) => {
 
   const handleLabelFormSubmit: TLabelFormProps['onSubmit'] = useCallback(async (data) => {
     const { text, reason, color, image } = data;
-    dispatch(personalActions.edit({ user, index, text, reason, color, image }));
+    const newImage = await submit(data);
+    const payload: IEditLabelPayload = { user, index, text, reason, color, image: newImage || image };
+    dispatch(personalActions.edit(payload));
     handleModalClose();
     // analytics
     gtag.event(EventAction.Edit, { event_category: EventCategory.Label, event_label: text });
-  }, [user, index, handleModalClose]);
+  }, [user, index, submit, handleModalClose]);
 
   return (
     <>
       <IconButton
-        className={className}
         icon={IconName.Pencil}
-        aria-label={TEXTS.BUTTON_TEXT_LABEL_EDIT}
-        data-tip={TEXTS.BUTTON_TEXT_LABEL_EDIT}
-        title={TEXTS.BUTTON_TEXT_LABEL_EDIT}
+        {...getElementLabelTipProps(TEXTS.BUTTON_TEXT_LABEL_EDIT)}
         onClick={handleClick}
+        {...otherProps}
       />
       <LabelFormModal
         open={open}
@@ -63,6 +66,7 @@ const EditLabelButton: React.FunctionComponent<TProps> = (props) => {
         label={label}
         escape={false}
         fragile={false}
+        loading={loading}
         onClose={handleModalClose}
         onSubmit={handleLabelFormSubmit}
       />