+++ src/components/LabelInfo/LabelInfo.tsx
@@ -1,15 +1,15 @@
 import classNames from 'classnames';
 import type React from 'react';
 import { memo, useMemo } from 'react';
+import * as TEXTS from '../../constants/texts';
 import { getShareURL } from '../../helpers/label';
 import type { IDataSet } from '../../models/DataSet';
 import type { ILabel } from '../../models/Label';
 import Subscription, { ISubscription } from '../../models/Subscription';
 import EditLabelButton from '../EditLabelButton/EditLabelButton';
-import Icon from '../Icon/Icon';
-import { IconName } from '../Icon/types';
 import LabelImageButton from '../LabelImageButton/LabelImageButton';
 import LabelSourceButton from '../LabelSourceButton/LabelSourceButton';
+import LabelSubscriptionButton from '../LabelSubscriptionButton/LabelSubscriptionButton';
 import RemoveLabelButton from '../RemoveLabelButton/RemoveLabelButton';
 import styles from './LabelInfo.module.scss';
 
@@ -34,31 +34,22 @@ const LabelInfo: React.FunctionComponent<TProps> = memo((props) => {
   const style: React.CSSProperties = useMemo(() => ({
     borderColor: color,
     boxShadow: color && `${color} 0 0 0.5rem inset`
-  }), [dataSet]);
-
-  const nameStyle: React.CSSProperties = useMemo(() => ({
-    color: (dataSet as ISubscription).color
-  }), [dataSet]);
+  }), [color]);
 
   return (
     <div className={classNames(className, styles.labelInfo)} style={style}>
-      {
-        isSubscriptionImplemented && (
-          <div className={styles.name}>
-            <Icon icon={IconName.InfoFill} />
-            <a href={dataSet.url} target="_blank" style={nameStyle}>
-              {dataSet.name}
-            </a>
-          </div>
-        )
-      }
-      {
-        label.reason && (
-          <div className={styles.reason}>
-            {label.reason}
-          </div>
-        )
-      }
+      <div
+        className={
+          classNames(
+            styles.reason,
+            {
+              [styles.empty]: !label.reason
+            }
+          )
+        }
+      >
+        {label.reason || TEXTS.LABEL_REASON_EMPTY_TEXT}
+      </div>
       <div className={styles.buttons}>
         {
           !isSubscriptionImplemented && (
@@ -79,17 +70,22 @@ const LabelInfo: React.FunctionComponent<TProps> = memo((props) => {
           label={label}
         />
         {
-          !isSubscriptionImplemented && (
+          !isSubscriptionImplemented ? (
             <RemoveLabelButton
               className={styles.button}
               user={user}
               index={index}
               label={label}
             />
+          ) : (
+            <LabelSubscriptionButton
+              className={styles.button}
+              subscription={dataSet}
+            />
           )
         }
       </div>
-    </div>
+    </div >
   );
 });