+++ src/components/RemoveLabelButton/RemoveLabelButton.tsx
@@ -3,6 +3,7 @@ import type React from 'react';
 import { useCallback } from 'react';
 import cache from '../../cache';
 import * as TEXTS from '../../constants/texts';
+import { getElementLabelTipProps } from '../../helpers/common';
 import * as gtag from '../../helpers/gtag';
 import type { ILabel } from '../../models/Label';
 import { actions as personalActions } from '../../store/slices/personal';
@@ -23,7 +24,7 @@ type TComponentProps = React.ComponentPropsWithoutRef<'button'>;
 type TProps = IProps & TComponentProps;
 
 const RemoveLabelButton: React.FunctionComponent<TProps> = (props) => {
-  const { className, user, index, label } = props;
+  const { user, index, label, ...otherProps } = props;
 
   const dispatch = useTypedDispatch();
 
@@ -41,12 +42,10 @@ const RemoveLabelButton: React.FunctionComponent<TProps> = (props) => {
 
   return (
     <IconButton
-      className={className}
       icon={IconName.DeleteForever}
-      aria-label={TEXTS.BUTTON_TEXT_LABEL_REMOVE}
-      data-tip={TEXTS.BUTTON_TEXT_LABEL_REMOVE}
-      title={TEXTS.BUTTON_TEXT_LABEL_REMOVE}
       onClick={handleClick}
+      {...getElementLabelTipProps(TEXTS.BUTTON_TEXT_LABEL_REMOVE)}
+      {...otherProps}
     />
   );
 };