+++ src/components/Settings/SubscriptionSection/ReloadSubscriptionButton/ReloadSubscriptionButton.tsx
@@ -2,6 +2,7 @@ import classNames from 'classnames';
 import type React from 'react';
 import { useCallback } from 'react';
 import * as TEXTS from '../../../../constants/texts';
+import { getElementLabelTipProps } from '../../../../helpers/common';
 import * as gtag from '../../../../helpers/gtag';
 import type { ISubscription } from '../../../../models/Subscription';
 import { actions as subscriptionsActions } from '../../../../store/slices/subscriptions';
@@ -41,9 +42,7 @@ const ReloadSubscriptionButton: React.FunctionComponent<TProps> = (props) => {
       className={classNames(className, lihkgCssClasses.settingOptionButton)}
       disabled={subscription.loading}
       icon={IconName.Refresh}
-      aria-label={TEXTS.BUTTON_TEXT_RELOAD_SUBSCRIPTION}
-      data-tip={TEXTS.BUTTON_TEXT_RELOAD_SUBSCRIPTION}
-      title={TEXTS.BUTTON_TEXT_RELOAD_SUBSCRIPTION}
+      {...getElementLabelTipProps(TEXTS.BUTTON_TEXT_RELOAD_SUBSCRIPTION)}
       onClick={handleClick}
     />
   );