+++ src/components/SettingsModalToggleButton/SettingsModalToggleButton.tsx
@@ -2,6 +2,7 @@ import type React from 'react';
 import { useCallback, useState } from 'react';
 import logo from '../../../assets/logos/libel.png';
 import { displayName } from '../../../package.json';
+import { getElementLabelTipProps } from '../../helpers/common';
 import IconButton from '../IconButton/IconButton';
 import SettingsModal from '../SettingsModal/SettingsModal';
 import styles from './SettingsModalToggleButton.module.scss';
@@ -26,9 +27,7 @@ const SettingsModalToggleButton: React.FunctionComponent = () => {
       <IconButton
         icon={icon}
         onClick={handleClick}
-        aria-label={displayName}
-        data-tip={displayName}
-        title={displayName}
+        {...getElementLabelTipProps(displayName)}
       />
       <SettingsModal
         open={open}