+++ src/constants/label.ts
@@ -4,4 +4,4 @@
  */
 export const DISPLAY_DATE_FORMAT = 'yyyy年MM月dd日 HH:mm:ss';
 
-export const SCREENSHOT_WIDTH = '600px';
+export const SCREENSHOT_WIDTH = 600;